requirements.txt
google-genai==1.16.1
fastapi==0.115.12
uvicorn==0.34.2
python-dotenv==1.1.0
aiohttp==3.11.18
aiogram==3.20.0